requirements.txt
# AI / LLM
anthropic>=0.18.0

# Configuration
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Excel / Report
openpyxl>=3.1.0

# Image Processing
Pillow>=10.0.0

# Logging
loguru>=0.7.0

# Async Support
aiohttp>=3.9.0
